Create Final.c
C program to Automate the Customer Shopping process in An Air Conditioning Store. It Comes with Options to Select the Energy Rating which is accompanied with Multiple Models for each Star Rating. The Process is Further Simplified by Adding Multiple Payment Methods which include Credit Card, Debit Card, Cash or EMI for the Financially Weak people.